+++ src/components/views/single-post/components/skeletons/interactions-skeleton.tsx
@@ -1,0 +1,8 @@
+export function InteractionsSkeleton() {
+  return (
+    <div className="flex flex-col gap-4">
+      <div className="h-10 w-full bg-gray-200 rounded-md" />
+      <div className="h-10 w-full bg-gray-200 rounded-md" />
+    </div>
+  )
+}