+++ src/components/views/single-post/post.client.tsx
@@ -4,16 +4,15 @@
 
 import { usePost } from "@/components/providers";
 import { Card } from "@/components/ui/card";
-import { PostActions } from "./components/post-actions";
-import { PostComments } from "./components/post-comments";
-import { PostLoading } from "./components/post-loading";
-import { PostNotFound } from "./components/post-not-found";
+import { PostActions } from "./components/ui/post-actions";
+import { PostComments } from "./components/ui/post-comments";
+import { PostLoading } from "./components/ui/post-loading";
+import { PostNotFound } from "./components/ui/post-not-found";
 
 export function PostViewClient() {
   const { post, isLoadingPost, postError } = usePost();
 
   if (isLoadingPost) {
-    return "TEGO NIE POWINNO BYC W OGOLE"
     return <PostLoading />;
   }