+++ src/components/views/single-post/post.server.tsx
@@ -1,12 +1,28 @@
+import { api } from "@/trpc/server";
+import { redirect } from "next/navigation";
+import { PostHeader } from "./components/ui/post-header";
+import { PostImage } from "./components/ui/post-image";
 import { Suspense } from "react";
-import { FirstLoadServer } from "./test_components/first-load.server";
+import { InteractionsServer } from "./components/interactions.server";
+import { InteractionsSkeleton } from "./components/skeletons/interactions-skeleton";
 
 export async function PostViewServer({ postId }: { postId: string }) {
+  const post = await api.post.getById({ id: postId });
+
+  if (!post) {
+    redirect("/")
+  }
+
   return (
-    <div className="flex flex-col gap-4 max-w-sm mx-auto">
-      <Suspense fallback={<div>Loading first load...</div>}>
-        <FirstLoadServer postId={postId} />
-      </Suspense>
-    </div >
-  );
+    <div className="max-w-sm mx-auto my-12">
+      <PostHeader post={post} />
+      <PostImage post={post} />
+
+      <div className="w-full">
+        <Suspense fallback={<InteractionsSkeleton />}>
+          <InteractionsServer postId={postId} initialPost={post} />
+        </Suspense>
+      </div>
+    </div>
+  )
 }